reorganize server unit tests
